requirements.txt
yt-dlp
pyyaml
customtkinter
spotdl
